requirements.txt
fastapi[standard]
sqlalchemy
pydantic-settings